requirements.txt
alembic==1.7.6
anyio==3.5.0
asgiref==3.5.0
asyncclick==8.0.3.2
asyncpg==0.25.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.15.0
click==8.0.3
databases==0.5.5
fastapi==0.73.0
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.0.1
passlib==1.7.4
psycopg==3.0.8
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-decouple==3.5
python-dotenv==0.19.2
PyYAML==5.4.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing_extensions==4.0.1
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
